+++ index.html
@@ -61,6 +61,6 @@
   </div>
 </body>
 
-<script src="script.js"></script>
+<script type="module" src="main.js"></script>
 
 </html>
